Significant change in database
- Added price to booking table
- Converted some integers to unsignedTinyInteger

+++ database/migrations/2022_11_30_114221_create_movies_table.php
@@ -26,7 +26,7 @@ return new class extends Migration
         'lang',
         ['ar', 'en', 'hu']
       );
-      $table->integer('duration');
+      $table->unsignedTinyInteger('duration');
       $table->decimal(
         'rating',
         total: 2,
